refactor(molecules): relocate filter and form components from organisms to molecules
- Move UnidadeFilters, InstituicaoFilter, GestoresEspaco, EspacoFormFields, ReservasFilters to presentation/molecules/
- Update GestoresEspaco to use UserAvatar Atom
- Delete unused HeaderSetor component
- Update all page and table import paths to reference new locations
- Verify tests and lint are fully passing
