+++ resources/js/presentation/pages/Reservas/ReservasPage.tsx
@@ -5,7 +5,7 @@ import { Head } from '@inertiajs/react';
 import { PlusCircle } from 'lucide-react';
 import { Suspense } from 'react';
 import { ReservasEmpty } from '@/presentation/molecules/ReservasEmpty';
-import { ReservasFilters } from '@/presentation/organisms/ReservasFilters';
+import { ReservasFilters } from '@/presentation/molecules/ReservasFilters';
 import { ReservasList } from '@/presentation/organisms/ReservasList';
 import { ReservasLoading } from '@/presentation/molecules/ReservasLoading';
